Revert "Add xattr support in windows"
This reverts commit 9c287cbe0d3b81ae59c2e9b17ca93b04d3779525.
