+++ internal/restic/node_netbsd.go
@@ -13,13 +13,18 @@ func (s statT) atim() syscall.Timespec { return s.Atimespec }
 func (s statT) mtim() syscall.Timespec { return s.Mtimespec }
 func (s statT) ctim() syscall.Timespec { return s.Ctimespec }
 
-// restoreExtendedAttributes is a no-op on netbsd.
-func (node Node) restoreExtendedAttributes(_ string) error {
-	return nil
+// Getxattr is a no-op on AIX.
+func Getxattr(path, name string) ([]byte, error) {
+	return nil, nil
 }
 
-// fillExtendedAttributes is a no-op on netbsd.
-func (node *Node) fillExtendedAttributes(_ string) error {
+// Listxattr is a no-op on AIX.
+func Listxattr(path string) ([]string, error) {
+	return nil, nil
+}
+
+// Setxattr is a no-op on AIX.
+func Setxattr(path, name string, data []byte) error {
 	return nil
 }
