+++ internal/restic/node_windows.go
@@ -105,80 +105,7 @@ func (s statT) mtim() syscall.Timespec {
 
 func (s statT) ctim() syscall.Timespec {
 	// Windows does not have the concept of a "change time" in the sense Unix uses it, so we're using the LastWriteTime here.
-	return s.mtim()
-}
-
-// restore extended attributes for windows
-func (node Node) restoreExtendedAttributes(path string) (err error) {
-	eas := []fs.ExtendedAttribute{}
-	for _, attr := range node.ExtendedAttributes {
-		extr := new(fs.ExtendedAttribute)
-		extr.Name = attr.Name
-		extr.Value = attr.Value
-		eas = append(eas, *extr)
-	}
-	if len(eas) > 0 {
-		if errExt := restoreExtendedAttributes(node.Type, path, eas); errExt != nil {
-			return errExt
-		}
-	}
-	return nil
-}
-
-// fill extended attributes in the node. This also includes the Generic attributes for windows.
-func (node *Node) fillExtendedAttributes(path string) (err error) {
-	var fileHandle windows.Handle
-
-	//Get file handle for file or dir
-	if node.Type == "file" {
-		if strings.HasSuffix(filepath.Clean(path), `\`) {
-			return nil
-		}
-		utf16Path := windows.StringToUTF16Ptr(path)
-		fileAccessRightReadWriteEA := (0x8 | 0x10)
-		fileHandle, err = windows.CreateFile(utf16Path, uint32(fileAccessRightReadWriteEA), 0, nil, windows.OPEN_EXISTING, windows.FILE_ATTRIBUTE_NORMAL, 0)
-	} else if node.Type == "dir" {
-		utf16Path := windows.StringToUTF16Ptr(path)
-		fileAccessRightReadWriteEA := (0x8 | 0x10)
-		fileHandle, err = windows.CreateFile(utf16Path, uint32(fileAccessRightReadWriteEA), 0, nil, windows.OPEN_EXISTING, windows.FILE_ATTRIBUTE_NORMAL|windows.FILE_FLAG_BACKUP_SEMANTICS, 0)
-	} else {
-		return nil
-	}
-	if err != nil {
-		err = errors.Errorf("open file failed for path: %s, with: %v", path, err)
-		return err
-	}
-	defer func() {
-		err := windows.CloseHandle(fileHandle)
-		if err != nil {
-			debug.Log("Error closing file handle for %s: %v\n", path, err)
-		}
-	}()
-
-	//Get the windows Extended Attributes using the file handle
-	extAtts, err := fs.GetFileEA(fileHandle)
-	debug.Log("fillExtendedAttributes(%v) %v", path, extAtts)
-	if err != nil {
-		debug.Log("open file failed for path: %s : %v", path, err)
-		return err
-	} else if len(extAtts) == 0 {
-		return nil
-	}
-
-	//Fill the ExtendedAttributes in the node using the name/value pairs in the windows EA
-	for _, attr := range extAtts {
-		if err != nil {
-			err = errors.Errorf("can not obtain extended attribute for path %v, attr: %v, err: %v\n,", path, attr, err)
-			continue
-		}
-		extendedAttr := ExtendedAttribute{
-			Name:  attr.Name,
-			Value: attr.Value,
-		}
-
-		node.ExtendedAttributes = append(node.ExtendedAttributes, extendedAttr)
-	}
-	return nil
+	return syscall.NsecToTimespec(s.LastWriteTime.Nanoseconds())
 }
 
 // restoreGenericAttributes restores generic attributes for Windows
